+++ Plugin.cs
@@ -31,7 +31,7 @@ namespace LOLItems
     {
         public const string GUID = "Ricky2148.etg.LOLItems";
         public const string NAME = "League of legends Items";
-        public const string VERSION = "3.1.1";
+        public const string VERSION = "3.2.0";
         public const string TEXT_COLOR = "#F1C232";
 
         internal static Harmony _Harmony;