+++ passive_items/EssenceReaver.cs
@@ -25,7 +25,7 @@ namespace LOLItems.passive_items
         public static void Init()
         {
             string itemName = ItemName;
-            string resourceName = "LOLItems/Resources/passive_item_sprites/sheen_pixelart_sprite";
+            string resourceName = "LOLItems/Resources/passive_item_sprites/essence_reaver_pixelart_sprite";
 
             GameObject obj = new GameObject(itemName);
 
@@ -34,8 +34,8 @@ namespace LOLItems.passive_items
             ItemBuilder.AddSpriteToObject(itemName, resourceName, obj);
 
             string shortDesc = "reaves essence";
-            string longDesc = "Grants Spellblade every few seconds. Spellblade: Empowers next bullet with some additional damage.\n\n" +
-                "\n";
+            string longDesc = "Increase damage\nGrants Spellblade every few seconds. Spellblade: Empowers next bullet with additional damage scaling with your stats. Restores a small amount of ammo each time spellblade is used.\n\n" +
+                "This magical reaver is made of self-forming ice, like Sheen. However, the higher quality magic imbued in the weapon overflows when the blade shatters, restoring slight amounts of ammunition to your weapons before reforming the blade.\n";
 
             ItemBuilder.SetupItem(item, shortDesc, longDesc, "LOLItems");